# 8310513: [s390x] Intrinsify recursive ObjectMonitor locking