s390 implementation of [JDK-8277180](https://bugs.openjdk.org/browse/JDK-8277180). PPC implementation for the same: https://github.com/openjdk/jdk/pull/7305

I had tested `tier1` on `fastdebug`, `release` vm. 

BenchMarking: 

```
./build/linux-s390x-server-release/jdk/bin/java -Xms4g -Xmx4g -jar dacapo-9.12-MR1-bach.jar h2 -s huge -t 1 -n 1

without patch: 
===== DaCapo 9.12-MR1 h2 PASSED in 223023 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 225686 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 219824 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 226719 msec =====



with patch: 
===== DaCapo 9.12-MR1 h2 PASSED in 167816 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 174368 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 170517 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 169349 msec =====
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8310513](https://bugs.openjdk.org/browse/JDK-8310513): [s390x] Intrinsify recursive ObjectMonitor locking (**Enhancement** - P4)


### Reviewers
 * [Lutz Schmidt](https://openjdk.org/census#lucy) (@RealLucy - **Reviewer**) ⚠️ Review applies to [9f48baa9](https://git.openjdk.org/jdk/pull/17975/files/9f48baa91028bd66dd8538b66401105c9c7a8ac5)
 * [Martin Doerr](https://openjdk.org/census#mdoerr) (@TheRealMDoerr - **Reviewer**) ⚠️ Review applies to [9f48baa9](https://git.openjdk.org/jdk/pull/17975/files/9f48baa91028bd66dd8538b66401105c9c7a8ac5)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/17975/head:pull/17975` \
`$ git checkout pull/17975`

Update a local copy of the PR: \
`$ git checkout pull/17975` \
`$ git pull https://git.openjdk.org/jdk.git pull/17975/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 17975`

View PR using the GUI difftool: \
`$ git pr show -t 17975`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/17975.diff">https://git.openjdk.org/jdk/pull/17975.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/17975#issuecomment-2038858492)